[libSyntax] Don't create deferred nodes for a libSyntax tree
Push the logic for creating deferred ParsedRawSyntaxNodes down to the
SyntaxParseActions. This allows us to handle deferred nodes more
efficiently in the SyntaxTreeCreator, which creates the libSyntax tree.
Instead of creating dedicated deferred nodes that are later converted
into recorded RawSyntax nodes, we directly record the RawSyntax nodes.
If the RawSyntax node is later requested to be recorded, we don't need
to do anything, but can just return the same node.

While we are leaking a small number of RawSyntax nodes this way, we have
a significant performance gain for libSyntax tree creation in deferred
mode. This is particularly important for the libSyntax parser transition
where the SyntaxParsingContext is always put into deferred mode to
discard all nodes after parsing, unless a SyntaxParseAction has been
explicitly requested.