[stdlib] Add dependency from static-executable-args.lnk to stdlib
The file static-executable-args.lnk was added to the stdlib install
component, but stdlib didn't depend on it, so when invoking
install-stdlib or install-swift-components, the file was not being
generated.

Because of the magic target marked as `ALL`, when one built all the
targets, and later issued a `install-*` target, the file was there by
a lucky coincidence.

The change only creates that dependency, so a call to install the stdlib
will also create this file.