build: add a workaound for the Windows build
The AST library for Swift exceeds the number of sections that the COFF
object file permits.  Add an explicit `/bigobj` flag to the build.

In theory, this flag should not be needed at all with D84038 applied to
llvm, which would add `/bigobj` unconditionally to all C/C++ objects
that are generated during the build.