[5.4] IRGen: We can only reuse methods for partial apply thunks if their abi uses swiftself for self
Explanation: A mismatch between the ABI of methods and the partial apply
thunk they are used for in an optimization -- the swiftself argument of
the closure context is is not reflected in the method's abi -- causes
crashes. Only certain methods use the swiftself register for the self
parameter. The code determining reuse of the method for the partial
apply thunk did not correctly compute the match.

Origination: This was introduced by an optimization in a commit in May.

Risk: Low. The fix uses the function that method signature lowering use
to determine whether swiftself is used in the method signature.

Testing: A regression test was added.

rdar://73777202