Work around a bug in the x86_64 swiftasync argument treatment.
r11 is the linker scratch register, so we need to make sure that
we don't end up in the linker when making cross-library direct
async calls.  Since there are currently only a few special
symbols that we do that for, we can achieve this by disabling
lazy binding for those symbols specifically.