Fix a Crash in Availability Checking of Defer Bodies
NIO exposed an issue in the new availability walkers where an implicit
check was not being performed. They were able to get this to crash by
using a defer statement - the body of which contains implicit
declarations that got run through the walker. This exposed a wider hole
in availability checking of defer statements. Namely, that it wasn't
happening.

This is a narrow fix (as opposed to the broader fix in
https://github.com/apple/swift/pull/36102) that is safer to take for
Swift 5.4.

rdar://74484150