Move getOwnershipKind from SILValue to ValueBase.
The API for values is on the ValueBase. SILValue is supposed to be a
pointer-like wrapper class. Accessing a value's API is always done as
'value->api()'. The ValueBase subclasses, like SingleValueInstruction,
need to inherit the API.

I didn't have time to fix all the cases of value.isOwnershipKind()
throughout the code.