Add static library configuration for _InternalSwiftSyntaxParser
When building command line tools that depend on SwiftSyntax, statically
linking this library can be preferred to produce a more portable binary
that doesn't directly depend on the system installation. Currently you
can achieve this benefit by shipping the dylib with your tool, and
editing the rpaths to prefer yours, but that's a lot of trouble.