[Deserialization] Output a diagnostic when deserializing an invalid decl
The previous change 47b068d4453776cd0feba754a4e41c416039ea26 was
partially reverted (when not allowing errors) as it output diagnostics
for valid code.

`Decl::isInvalid` had a special case for `AccessorDecl` where if its
interface type hadn't already been computed, it would fallback to
checking whether its storage (ie. its `VarDecl`) was valid. Since the
`VarDecl` isn't fully deserialized when deserializing the
`AccessorDecl`, this sometimes causes `isInvalid` to be true, even
though it isn't actually invalid.

The fallback introduced to `isInvalid` was to prevent a cycle during
typechecking caused by `SimpleDidSetRequest` - during `didSet`
typechecking, the body would be typechecked, which would then check
whether the declaration is valid, calling the typechecking for the
declaration again (and so on).

This change removes the special case from `isInvalid`. The result of
`isInvalid` is then correct for `AccessorDecl`, preventing spurious
diagnostics.

The `isInvalid` check in `performAbstractFuncDeclDiagnostics` has been
removed to prevent the cycle there - it works perfectly fine without it.
The unused variable diagnostics are skipped on seeing an error in the
body anyway. The opaque type matching diagnostics make sense to keep
when the return type is valid.

Removing that check exposes another cycle when computing captures in
body typechecking caused by `shouldMarkAsObjC` checking `isInvalid` as
well. It already checks for `didSet`, so have moved that check above the
`isInvalid` check.

Resolves rdar://74541834