GSB: Replace RequirementSource::Derived with ::Layout
'Derived' was not a great name, since we already use the term
'derived requirement source' to mean something else.

A Derived source was only added in one place, when recording
a superclass constraint -- the idea is that this source
supercedes any explicit layout constraint, eg

    class SomeClass {}
    func foo<T>(_: T) where T : SomeClass, T : AnyObject {}

Here we have two sources for the 'T : AnyObject' layout constraint:

    Explicit: T
    Explicit: T -> Derived

Note that the 'Derived' requirement source does not store
a 'proof' -- we can't figure out _how_ we determined that
the explicit 'T : AnyObject' constraint is redundant here.

In the case where a superclass requirement makes a protocol
conformance redundant, we do have a 'proof', because the
'Superclass' requirement source stores a conformance:

    class SomeClass : SomeProto {}
    func foo<T>(_: T) where T : SomeClass, T : SomeProto {}

    Explicit: T
    Explicit: T -> Superclass: [SomeClass : P]

From looking at the second requirement source, we can
determine that the requirement was imposed by the explicit
constraint 'T : SomeClass'.

For the 'Layout' requirement source, there's not really a
"conformance", so we can just store the superclass type:

    Explicit: T
    Explicit: T -> Layout: SomeClass